Add support for OS X
Because OS X doesn’t read .bashrc, I’ve created a .bash_profile that
checks if we’re in OS X and source all the relevant files if we are.
These files include the git autocomplete and __git_ps1 for bash prompt.
Of course the .bashrc file is also sourced.
If we’re not in OS X then hopefully it will just skip this, source the
.bashrc as expected and work the same way.